Dramatically relax rules for key names.
SD/Prophet need to be able to have spaces in key names for storing
aliases, and ., :, and / characters for automatic addition of sources
for friendly names.  There's little reason not to allow any characters
that won't screw up the parsing except for "because git does it that
way". Dot characters in variable names do allow a chance for identifier
collision in some very weird cases, but it's doubtful that this will
cause problems in practice.