Keep track of which file each key was loaded from.
Ow. We need this because otherwise, when loading, say,
a user config file that contains a value that's also
contained in a global config file that has already
been loaded, instead of overriding the old key with
the one from the new file, it would treat the key
has having multiple values. Which isn't the desired
behaviour.