Add the "encoding" attribute and use it.
To encode and decode config files. Necessary because, if no encoding is
specified, and a UTF-8 string is saved to a config file, when it is read back
in, it will not be properly read as UTF-8. The same would go for any coding
other than Latin-1 (maybe).

I'm not sure what encoding Git assumes; perhaps the locale-specified encoding.
If so, perhaps we should set the file layters to `:locale` if no encoding
is specified?