Enforce that all keys must have sections
Previously, git correctly parsed configuration variables before the
first section marker, although it coult not set or unset such keys.  In
git 1.7.4.4 and later, the code was modified to match the documentation,
and this misfeature was removed.

Alter the parsing code and methods to enforce this restriction in the
module as well, for consistency with git.