Track and allow lookup of the case that keys were defined in
Git canonicalizes the section and key name into lower case for all
lookups; subsection names are case-sensitive.  However, there are times
when is is useful to be able to determine the original case used to set
the key.

Store the original case, and provide a 'original_case' method to
retrieve the kase used for a given key.  This also alters the 'name'
argument to the 'callback' of 'parse_content' to be the original case,
and not the canonical lower case.