Store the cleaned-up filename path, with ../s resolved
Cwd::abs_path examines the filesystem to resolve /foo/../bar/ to /bar --
while taking into account symlinks and other complicating factors.  Use
it to store the simplest form of the filename loaded.

As Cwd::abs_path returns undef if it cannot resolve the path, default
back to the passed-in value, for error messages and the like.