dir_file() does not work as documented
The documentation claims it returned ".git/config", but it actually
returns ".gitconfig". This is particularly inconvenient when doing a
read/modify/write of one and the same configuration in place. The
filename parameter to set() is required, but the only way to get the
name of the config file which was read before (the dir_file() method)
returns the wrong file name.