Return a more correct `dir_file` for Config::GitLike::Git
Previously, the documentation claimed it returned `.git/config`, but
actually returned `.gitconfig`. This is particularly inconvenient when
doing a read/modify/write of a file in-place.

Return `.git/config` from `dir_file` if called with no argument.  This
change is not backwards-compatible, but is more correct to the
expected use, and the documentation.

If passed an optional directory path, `dir_file` now returns the
fully-qualified path to the config file that would have been loaded
from; this allows it to be used in bare repository directories
correctly.

Fixes #14.